# api_spec/openapi.yaml
openapi: 3.0.3
info:
  title: Media Server Pro API
  description: >
    Modular media streaming server providing video/audio streaming with HLS
    adaptive streaming, thumbnail generation, analytics, user authentication,
    mature content scanning, and admin management.
  version: 0.0.3
  license:
    name: Proprietary
servers:
  - url: /
    description: Same-origin server
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
  schemas:
    APIEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          description: Response payload (type varies by endpoint)
        error:
          type: string
      required: [ success ]
...
    MediaItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [ video, audio, unknown ]
        size:
          type: integer
          format: int64
        duration:
          type: number
        width:
          type: integer
        height:
          type: integer
        bitrate:
          type: integer
          format: int64
        codec:
          type: string
        container:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        thumbnail_url:
          type: string
        date_added:
          type: string
          format: date-time
        date_modified:
          type: string
          format: date-time
        views:
          type: integer
        last_played:
          type: string
          format: date-time
        is_mature:
          type: boolean
        mature_score:
          type: number
        blur_hash:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
    MediaListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MediaItem'
        total_items:
          type: integer
        total_pages:
          type: integer
        scanning:
          type: boolean
        user_ratings:
          type: object
          additionalProperties:
            type: number
            format: float
          description: Map of media_id to user rating
        initializing:
          type: boolean
          description: True when media library is still loading
...
paths:
...
  /media:
    get:
      tags: [ streaming ]
      summary: Stream a media file
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
        - name: quality
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Media stream with Range support
        '206':
          description: Partial content
  /download:
    get:
      tags: [ streaming ]
      summary: Download a media file
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File download
...
  /thumbnail:
    get:
      tags: [ thumbnails ]
      summary: Get thumbnail for media item
      parameters:
        - name: id
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [ placeholder, audio_placeholder, censored ]
      responses:
        '200':
          description: JPEG image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
...
  /api/media:
    get:
      tags: [ media ]
      summary: List media items
      parameters:
        - name: sort
          in: query
          schema:
            type: string
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [ asc, desc ]
        - name: limit
          in: query
          schema:
            type: integer
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            maximum: 50000
        - name: type
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: tags
          in: query
          schema:
            type: string
          description: Comma-separated tags
        - name: is_mature
          in: query
          schema:
            type: string
            enum: [ 'true', 'false' ]
      responses:
        '200':
          description: Media list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MediaListResponse'
...
  /api/admin/media:
    get:
      tags: [ admin-media ]
      summary: List media (admin)
      security:
        - sessionCookie: [ ]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Media items
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/MediaItem'
...